Moved the "gutter" into the gallery gallery component.
The gutter shouldn't be included in the actual width of the gallery element, because that has a knock on effect on the body and messes up the navbar on mobile.
So the gutter has been moved into the gallery component and is now only used in the layout algorithm.
The gallery also hides it's x overflow now to prevent the same knockon effect on the body.